requirements.txt
# Requirements for Brain Tumor Classification Gradio App

# Web interface framework
gradio>=4.0.0

# Deep Learning framework
tensorflow>=2.13.0

# Image processing
Pillow>=9.5.0

# Numerical computing
numpy>=1.24.0